Update cam.py
Changed things up a bit a bit to steamline things. This now uses pigpio to find the daemon instead of OS. Has a problem of taking forever because it gets stuck in loop of rotating motor. Will fix soon, either by multithreading or pushing that onto the Pi (definitely the latter since it'll be less intensive on the computer's CPU)